Changed the way the mod works fundamentally
Changed the functionality over from replacing the edges with "empty" models (which resulted in shadows being cast from the invisbile models) to a script which dynamically removes edges on all armor blocks. This doesn't change anything in the effect of the mod, but will hopefully alleviate some of its quirks.

Fixes #3 .