Base-level change in how the mod works
Completely changed around the way the mod works: Instead of changing the planet definitions as a whole, the mod now mainly implements its changes via a script that runs once and changes only the ore mappings on planets and the ores that spawn on asteroids. This increases compatability with other mods changing planets dramatically and also makes it very easy to include support for custom planets.
The script was written by Meridius and adjusted by me.